requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
packaging==24.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-gitlab==4.11.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.1.0
urllib3==2.2.3